fix(security): CodeQL 경고 8건 처리 — 오픈 리다이렉트 · 페이지 오버플로 500 등
- 로그인 next: URL 로 해석해 같은 출처일 때만 이동 ('/\evil.example' 우회 차단)
- page × size 오버플로(음수 OFFSET → 500) → page ≤ 100,000 검증 + long 계산, ApiIT 회귀 테스트
- 감사 실패 로그 줄바꿈 제거 · 임시 디렉터리 소유자 전용 · URL 상태는 알려진 키만
- 경로 파라미터 형식 검사 · API 클라이언트가 '..' '//' '\' 경로 거부
- VERIFICATION 38~40

+++ app/src/main/java/com/buildrisk/radar/api/AlertQueryService.java
@@ -50,7 +50,7 @@ public class AlertQueryService {
                         + "LIMIT :lim OFFSET :off")
                 .param("tt", CompanyQueryService.blank(targetType)).param("sev", CompanyQueryService.blank(severity))
                 .param("st", st).param("since", since).param("tk", CompanyQueryService.blank(targetKey))
-                .param("rc", CompanyQueryService.blank(ruleCode)).param("lim", size).param("off", page * size)
+                .param("rc", CompanyQueryService.blank(ruleCode)).param("lim", size).param("off", (long) page * size)
                 .query((rs, i) -> new AlertRow(rs.getLong(1), rs.getString(2), rs.getInt(3), rs.getString(4),
                         rs.getString(5), rs.getString(6), rs.getString(7), rs.getString(8), rs.getString(9),
                         rs.getString(10), rs.getString(11), rs.getString(12), rs.getObject(13, OffsetDateTime.class),
